Simplify starting a plan
Previously we were figuring out which exercise
was just entered the last time this day the plan
was opened. This caused our current faster method
of opening the exercises immediately to have a
delay, where the user would see the first selected,
and then the most recent one selected.
Rather than have all this logic we now have a
quicker setup that retrieves the first one instead.