Quit deploying if there are unstaged changes
If the source code is edited while deploy.sh is
running, there is a chance the APK released
and the APK associated with the tag are not
the same. This has caused a few f-droid
reproducible build failures.

We should be able to check by using git diff-files,
because GitHub is creating the tag on the latest
commit.