Increase versionCode on universal releases
If you aren't a developer this change doesn't matter.
For developers, this means when you install the app
on your phone in debug/release from your PC,
it won't clash with the f-droid version. Previously
if you had installed the app manually then f-droid
would consider it "outdated" and suggest you
update.